requirements.txt
appdirs==1.4.3
attrs==19.3.0
cached-property==1.5.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
idna==2.8
isodate==0.6.0
lxml==4.4.1
pycparser==2.19
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.9.1
six==1.13.0
urllib3==1.25.7
zeep==3.4.0
